send (stream) data blocked when flow control limit doesn't allow to send any more data

+++ src/main/java/net/luminis/quic/frame/DataBlockedFrame.java
@@ -37,7 +37,7 @@ public class DataBlockedFrame extends QuicFrame {
     public DataBlockedFrame() {
     }
 
-    public DataBlockedFrame(int streamDataLimit) {
+    public DataBlockedFrame(long streamDataLimit) {
         this.streamDataLimit = streamDataLimit;
     }
